+++ ICLCore/src/iclPoint.cpp
@@ -5,6 +5,6 @@ namespace icl{
   const Point Point::null(0,0);
 
   float Point::distanceTo(const Point &p) const{
-    return sqrt(pow((p.x-x),2)+pow(p.y-y,2));
+    return sqrt(pow((float) (p.x-x), 2) + pow((float) (p.y-y), 2));
   }
 }
